requirements.txt
allure-pytest==2.13.5
attrs==25.3.0
build==1.2.2.post1
cfgv==3.4.0
chardet==5.2.0
click==8.1.8
contourpy==1.3.1
cycler==0.12.1
distlib==0.3.9
et_xmlfile==2.0.0
execnet==2.1.1
filelock==3.18.0
fonttools==4.56.0
gherkin-official==29.0.0
identify==2.6.9
iniconfig==2.1.0
kiwisolver==1.4.8
Mako==1.3.9
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
mdurl==0.1.2
nodeenv==1.9.1
numpy==2.2.4
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
parse==1.20.2
parse_type==0.6.4
pillow==11.1.0
pip-tools==7.4.1
platformdirs==4.3.7
pluggy==1.5.0
pre_commit==4.2.0
psutil==7.0.0
Pygments==2.19.1
pyparsing==3.2.3
pyproject_hooks==1.2.0
pytest==8.3.5
pytest-bdd==8.1.0
pytest-xdist==3.6.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
rich==13.9.4
setuptools==78.0.2
six==1.17.0
typing_extensions==4.12.2
tzdata==2025.2
virtualenv==20.29.3
wheel==0.45.1
xlrd==2.0.1
